+++ tomcat/src/main/java/com/techcourse/controller/UserRegisterController.java
@@ -1,11 +1,9 @@
 package com.techcourse.controller;
 
 import com.techcourse.service.UserService;
-import java.io.IOException;
 import org.apache.coyote.http11.handler.AbstractController;
-import org.apache.coyote.http11.handler.controllerResponse.ApplicationResponse;
+import org.apache.coyote.http11.handler.applicationResponse.ApplicationResponse;
 import org.apache.coyote.http11.httpRequest.HttpRequest;
-import org.apache.coyote.http11.httpResponse.HttpResponse;
 
 public class UserRegisterController extends AbstractController {
