+++ tomcat/src/main/java/com/techcourse/service/UserService.java
@@ -8,9 +8,9 @@ import org.apache.catalina.session.Session;
 import org.apache.catalina.session.SessionManager;
 import org.apache.coyote.http11.general.CommonHeaderKeys;
 import org.apache.coyote.http11.general.Cookies;
-import org.apache.coyote.http11.handler.controllerResponse.ApplicationResponse;
-import org.apache.coyote.http11.handler.controllerResponse.JsonResponse;
-import org.apache.coyote.http11.handler.controllerResponse.StaticFileResponse;
+import org.apache.coyote.http11.handler.applicationResponse.ApplicationResponse;
+import org.apache.coyote.http11.handler.applicationResponse.JsonResponse;
+import org.apache.coyote.http11.handler.applicationResponse.StaticFileResponse;
 import org.apache.coyote.http11.general.CookieParser;
 import org.apache.coyote.http11.httpRequest.HttpRequest;
 import org.apache.coyote.http11.httpResponse.HttpStatus;
