+++ tomcat/src/main/java/org/apache/coyote/http11/handler/Controller.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.handler;
 
-import org.apache.coyote.http11.handler.controllerResponse.ApplicationResponse;
+import org.apache.coyote.http11.handler.applicationResponse.ApplicationResponse;
 import org.apache.coyote.http11.httpRequest.HttpRequest;
 
 public interface Controller {
